เพิ่ม icon / แก้ fragment
