--- app/src/main/java/com/android/orc/ocrapplication/dashboard/fragment/GalleryFragment.java
@@ -1,70 +1,0 @@
-package com.android.orc.ocrapplication.dashboard.fragment;
-
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.android.orc.ocrapplication.R;
-
-
-/**
- * Created by j.poobest on 9/24/2017 AD.
- */
-
-public class GalleryFragment extends Fragment {
-
-    public GalleryFragment() {
-        super();
-    }
-
-    public static GalleryFragment newInstance() {
-        GalleryFragment fragment = new GalleryFragment();
-        Bundle args = new Bundle();
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        init(savedInstanceState);
-
-        if (savedInstanceState != null)
-            onRestoreInstanceState(savedInstanceState);
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_gallery_dashboard, container, false);
-        initInstances(rootView, savedInstanceState);
-        return rootView;
-    }
-
-    @SuppressWarnings("UnusedParameters")
-    private void init(Bundle savedInstanceState) {
-        // Init Fragment level's variable(s) here
-    }
-
-    @SuppressWarnings("UnusedParameters")
-    private void initInstances(View rootView, Bundle savedInstanceState) {
-        // Init 'View' instance(s) with rootView.findViewById here
-        // Note: State of variable initialized here could not be saved
-        //       in onSavedInstanceState
-    }
-
-    @Override
-    public void onSaveInstanceState(Bundle outState) {
-        super.onSaveInstanceState(outState);
-        // Save Instance (Fragment level's variables) State here
-    }
-
-    @SuppressWarnings("UnusedParameters")
-    private void onRestoreInstanceState(Bundle savedInstanceState) {
-        // Restore Instance (Fragment level's variables) State here
-    }
-
-}
