+++ app/src/main/res/menu/menu_bottom_navigation.xml
@@ -11,7 +11,7 @@
     <item
         android:id="@+id/action_gallery"
         android:checked="false"
-        android:icon="@drawable/ic_bottom_navigation_gallery"
-        android:title="gallery"
+        android:icon="@drawable/ic_bottom_navigation_favorite"
+        android:title="favorite"
         app:showAsAction="ifRoom" />
 </menu>
